fix certification error issue
The default make-cert script will create cert file only containing k8s
master IP. However, in our env which is built on top of openstack VMs,
the external IP is not included. Then when client accessing the K8S
using external IP, the certification will fail.

This is not an final fix to this issue as we need to change IP for every
setup. So it works like an tomb for reminding only.